requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
Django==2.1.8
django-countries==5.3.3
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==3.0.8
django-extra-views==0.13.0
django-haystack==3.0b2
django-oscar==2.1
django-oscar-api==2.0.2
django-phonenumber-field==3.0.1
django-restframework==0.0.1
django-tables2==2.2.1
django-treebeard==4.3.1
django-urls==1.1.3
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
easy-thumbnails==2.7
factory-boy==2.12.0
Faker==4.1.2
flake8==3.8.3
importlib-metadata==1.7.0
mccabe==0.6.1
phonenumbers==8.12.9
Pillow==7.2.0
purl==1.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.2.0
Whoosh==2.7.4
zipp==3.1.0
